Fix #489 — Split oversized spec files to stay under 300-line limit
